feat: home page design

+++ src/features/project/auth-store.ts
@@ -8,6 +8,7 @@ interface AuthInitialProps {
   isAuthenticated: boolean;
   user: UserResponse | null;
   setIsAuthenticated: (value: boolean) => void;
+  loadUser: () => Promise<void>;
   setUser: (user: UserResponse) => void;
   logout: () => Promise<void>;
 }
@@ -19,12 +20,32 @@ const useAuthStore = create<AuthInitialProps>((set) => ({
   setAccessToken: (token: string) => set({ accessToken: token }),
   setIsAuthenticated: (value: boolean) => set({ isAuthenticated: value }),
   setUser: (user: UserResponse) => set({ user }),
+  loadUser: async () => {
+    const {
+      data: { session },
+      error: errorSession,
+    } = await supabase.auth.getSession();
+
+    if (!errorSession && session) {
+      set({ user: session?.user as UserResponse, isAuthenticated: true, accessToken: session?.access_token });
+    } else {
+      const { error: errorLogout } = await supabase.auth.signOut();
+
+      if (!errorLogout) {
+        set({ user: null, isAuthenticated: false, accessToken: null });
+        window.localStorage.clear();
+        window.location.replace('/sign-in');
+      }
+    }
+  },
   logout: async () => {
     const { error } = await supabase.auth.signOut();
 
-    if (!error) set({ user: null, isAuthenticated: false });
-    window.localStorage.clear();
-    window.location.replace('/sign-in');
+    if (!error) {
+      set({ user: null, isAuthenticated: false, accessToken: null });
+      window.localStorage.clear();
+      window.location.replace('/sign-in');
+    }
   },
 }));
